C-S_ver_xx/C_S_ver_0.5 _PSQL/requirements.txt
psycopg2-binary==2.9.7
typing_extensions>=4.0.0
